System Mini games part1
Everything seems to work, there is the sequencing system for the mini games, you make a mistake in one and it goes to the end, it already indicates on the console when you have finished the mini games.

You have to implement the provisional interfaces to have the complete experience, do the analysis and see what follows.